refactor(odata): move the paging wrapper into pagination.rs (readability ratchet)
The post-merge mod.rs was 550 lines (my ~85-line read_entity_set_from_query_plane
wrapper on top of the merged ARN-68 gap-reconcile block), crossing the 500-line
readability ratchet gate (PROD_FILES_GT500 83 -> 84). Relocate the wrapper to the
pagination module where it belongs — it is the pagination layer — and re-export it.
mod.rs 452, pagination.rs 387; ratchet back to baseline (GT500=83). Behavior
unchanged; query_plane_read suite 47/47.
